Revamped crafting image preview.
Added crafted item name in the crafting meter with overflow control.

Refined styles and moved the roster out of the facilities panel grid.
